Programmatically instantiate BST.
Still figuring out how I want to do this in Unity. Originally this was a
WPF program and I got frustrated with WPF, so I thought I'd switch to
Unity. The only downside is that I don't really know how to do things
"the right way" in Unity so this is probably going to be ugly.

My goal here is to use Unity to visually represent a BST and possibly
operations on the BST (if I find time to animate it). The main challenge
so far is creating a clear separation between my generic BST and Unity.
I don't want the Unity API to pollute my BST classes. So far, I'm
thinking this will work as follows:

- Upon instantiation, create the BST with a specific implementation,
then instruct Unity to read the BST and create the visual nodes

- When the BST is modified, call the relevant operation on the BST and
then call an update method which will cause Unity to scan the BST and
update the visuals as necessary.

This will create a fundamental separation between Unity and the BST.